% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{aggregate_preds}
\alias{aggregate_preds}
\title{Aggregate future predictions}
\usage{
aggregate_preds(species)
}
\arguments{
\item{species}{character Species name, as specified when fitting Maxent models.}
}
\value{
A RasterLayer with the predicted average suitability per species.
}
\description{
Aggregate future predictions
}
