% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{combine_pred}
\alias{combine_pred}
\title{Combine future predictions from a Maxent model}
\usage{
combine_pred(model, scenario)
}
\arguments{
\item{model}{A maxent model, as created by dismo.}

\item{scenario}{Character. Either "rcp45" or "rcp85".}
}
\value{
A rasterstack.
}
\description{
Combine future predictions from a Maxent model
}
