% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compare_suitab_futu_pres}
\alias{compare_suitab_futu_pres}
\title{Compare future vs present suitability per species}
\usage{
compare_suitab_futu_pres(futu.suitab)
}
\arguments{
\item{futu.suitab}{RasterStack (6 layers) containing future suitabilities per species.}
}
\value{
A plot, and a RasterStack with the difference in suitabilities (future - present)
}
\description{
Compare future vs present suitability per species
}
