% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{crop_bioregions}
\alias{crop_bioregions}
\title{Crop raster by bioregions}
\usage{
crop_bioregions(ras)
}
\arguments{
\item{ras}{Raster* object}
}
\value{
Cropped Raster* object.
}
\description{
Crop raster by bioregions
}
