% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_packages.R
\name{delete_packages}
\alias{delete_packages}
\title{Delete packages library}
\usage{
delete_packages()
}
\description{
Delete special packages library created by \code{\link{recreate_packages}}: "~/.checkpoint".
}
