% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ensemble_mean}
\alias{ensemble_mean}
\title{Aggregate future predictions: Calculate ensemble mean}
\usage{
ensemble_mean(species, scenario)
}
\arguments{
\item{species}{character Species name, as specified when fitting Maxent models.}

\item{scenario}{Character. Either "rcp45" or "rcp85".}
}
\value{
A RasterLayer with the predicted average suitability per species.
}
\description{
Aggregate future predictions: Calculate ensemble mean
}
