% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot6maps}
\alias{plot6maps}
\title{Plotting function: six maps}
\usage{
plot6maps(ras, maptype)
}
\arguments{
\item{ras}{RasterStack or RasterBrick with 6 layers (one per species, plus all species).}

\item{maptype}{Character. Either "suitab" for suitability values between 0 and 1, "sd" for plotting standard deviations (between 0 and 0.5), or "diff" for plotting differences in suitability (between -0.5 and +0.5).}
}
\value{
A plot
}
\description{
Plotting function: six maps
}
