% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_ensemble}
\alias{plot_ensemble}
\title{Plot ensemble mean or standard deviation}
\usage{
plot_ensemble(mean.sd, scenario)
}
\arguments{
\item{mean.sd}{character Either "mean" or "sd" to return the ensemble mean or standard deviation, respectively.}

\item{scenario}{Character. Either "rcp45" or "rcp85".}
}
\value{
A RasterStack and a plot.
}
\description{
Plot ensemble mean or standard deviation
}
