% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_futclim}
\alias{read_futclim}
\title{Read future climate layers}
\usage{
read_futclim(gcmdata)
}
\arguments{
\item{gcmdata}{character Model and scenario (e.g. "miroc5_rcp45_bio_2050").}
}
\value{
A raster stack, cropped by bioregions.
}
\description{
Read future climate layers
}
\details{
Note path to GIS layers is hard-coded.
}
