% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_pres_suitab}
\alias{read_pres_suitab}
\title{Read present suitability raster for all species}
\usage{
read_pres_suitab()
}
\value{
A RasterStack with the present potential distribution of all species according to Maxent models.
}
\description{
Read present suitability raster for all species
}
