% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_presclim}
\alias{read_presclim}
\title{Load present climate}
\usage{
read_presclim()
}
\value{
A raster stack, cropped by bioregions.
}
\description{
Load current climate layers from my hard drive (note path is hard-coded).
}
