% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_packages.R
\name{recreate_packages}
\alias{recreate_packages}
\title{Recreate packages}
\usage{
recreate_packages(date = "2017-09-19")
}
\arguments{
\item{date}{character. Date ("YYYY-MM-DD" format). This will be used by \code{checkpoint} to download the corresponding package versions from CRAN for that date. Packages will be downloaded to a local folder ("~/.checkpoint) and these specific versions will be used for the session (Note that restarting the session will reset to the normal user library).}
}
\description{
Recreate dependencies (packages) used in the project as they were in a specified date.
}
