% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Andes and Nearctic biogeographic regions.}
\format{A SpatialPolygonsDataFrame with 2 polygons defining the limits of the Nearctic and Andean region.}
\usage{
regions
}
\description{
Andes and Nearctic biogeographic regions, as defined in Morrone (2002).
}
\references{
Morrone, J.J. 2002. Biogeographical regions under track and cladistic scrutiny. Journal of Biogeography 29: 149 – 152.
}
\keyword{datasets}
